Deploy-Agent/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
loguru
bcrypt
pyjwt
psutil
py-cpuinfo
jinja2
httpx
pyyaml